# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RentKarRideKar API Gateway",
    "description": "Complete interactive API documentation for RentKarRideKar services including Auth, Vehicle, KYC, Order, Earnings, Referral, Wallet, Disputes, Ratings, Communication, and Admin services.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/kyc/upload": {
      "post": {
        "summary": "Upload KYC Document",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DRIVER",
                "VEHICLE_OWNER"
              ]
            }
          },
          {
            "name": "documentType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "AADHAAR_FRONT",
                "AADHAAR_BACK",
                "PAN_CARD",
                "DRIVING_LICENSE",
                "SELFIE",
                "RC_BOOK",
                "INSURANCE",
                "PUC_CERTIFICATE",
                "FITNESS_CERT",
                "PERMIT",
                "COMMERCIAL_PERMIT"
              ]
            }
          },
          {
            "name": "vehicleId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded successfully"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
    "/kyc/status": {
      "get": {
        "summary": "Get Personal KYC Status",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
    "/kyc/status/{vehicleId}": {
      "get": {
        "summary": "Get Vehicle KYC Status",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
    "/kyc/vehicle/{vehicleId}": {
      "get": {
        "summary": "Get Vehicle Docs",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
    "/kyc/submit": {
      "patch": {
        "summary": "Submit User KYC",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "DRIVER",
                      "VEHICLE_OWNER"
                    ]
                  }
                },
                "required": [
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submitted successfully"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
    "/kyc/submit-vehicle": {
      "patch": {
        "summary": "Submit Vehicle Docs for KYC",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "example": "VEHICLE_OWNER"
                  }
                },
                "required": [
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
...
    "/kyc/admin/documents/{id}": {
      "patch": {
        "summary": "Approve or Reject document",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "APPROVED",
                      "REJECTED"
                    ]
                  },
                  "rejectionReason": {
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
    "/kyc/admin/verifications/{id}/reject": {
      "patch": {
        "summary": "Reject Overall KYC Verification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rejectionReason": {
                    "type": "string"
                  }
                },
                "required": [
                  "rejectionReason"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
    "/kyc/admin/verifications/{id}/approve": {
      "patch": {
        "summary": "Approve Overall KYC Verification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "KYC & Verification Service"
        ]
      }
    },
...
  }
}